requirements.txt
openai-whisper
moviepy
pytest
pytest-cov
torch
torchaudio
soundfile
transformers
openai
python-dotenv
rich